server/requirements.txt
Flask==3.1.2
mediapipe==0.10.14
opencv-python==4.11.0.86
Pillow==11.3.0
Werkzeug==3.1.3